requirements.txt
fastapi
uvicorn
python-multipart
torch
torchvision
pillow
numpy
scikit-learn
matplotlib
jinja2
python-dotenv
streamlit
requests
